+++ src/arch/x86_64/mm/include/vmm.h
@@ -65,16 +65,16 @@ extern "C" {
 #define OFFSET_INDEX(x) ( ( x ) & 0x0FFF )
 
 // 页目录数据类型
-typedef uint32_t pgd_t;
+typedef uint64_t pgd_t;
 
 // 页表数据类型
-typedef uint32_t pte_t;
+typedef uint64_t pte_t;
 
 // 页表成员数
 #define PGD_SIZE ( PAGE_SIZE / sizeof( pte_t ) )
 
 // 页表成员数
-#define PTE_SIZE ( PAGE_SIZE / sizeof( uint32_t ) )
+#define PTE_SIZE ( PAGE_SIZE / sizeof( uint64_t ) )
 
 // 映射 512MB 内存所需要的页表数
 #define PTE_COUNT 128
